requirements.txt
pandas
numpy
matplotlib
scipy
hydra-core